Add google test

+++ List3/ex2Functions.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
-int multiply (int multiplier, int a);    // Function for multiplying value "a" 
+float multiply (float multiplier, float a);    // Function for multiplying value "a" 
 
-int multiply (int multiplier, int a)
+float multiply (float multiplier, float a)
 {
     return a *= multiplier;
 }
